refactor: Move needs-assessment data sourcing helpers into their own util file.
